refactor(!): remove UI
removes messaging, router, relocates UI controller to top-level, re-introduces sendToPropertyInspector, re-locates current PI (now action) to UI controller.
