+++ src/common/logging/index.ts
@@ -2,5 +2,4 @@ export * from "./console-target";
 export * from "./format";
 export * from "./level";
 export * from "./logger";
-export * from "./routing";
 export * from "./target";
