--- src/common/messaging/index.ts
@@ -1,3 +1,0 @@
-export * from "./gateway";
-export { type StatusCode } from "./message";
-export { MessageResponder } from "./responder";
