+++ src/plugin/__tests__/index.test.ts
@@ -6,7 +6,6 @@ import { SingletonAction } from "../actions/singleton-action";
 import { connection } from "../connection";
 import streamDeckAsDefaultExport, { streamDeck } from "../index";
 import { logger } from "../logging";
-import { route } from "../ui/route";
 
 jest.mock("../../common/i18n");
 jest.mock("../logging");
@@ -76,7 +75,6 @@ describe("index", () => {
 		expect(index.LogLevel).toBe(LogLevel);
 		expect(index.SingletonAction).toBe(SingletonAction);
 		expect(index.Target).toBe(Target);
-		expect(index.route).toBe(route);
 	});
 
 	/**
