--- src/plugin/ui/__mocks__/router.ts
@@ -1,2 +1,0 @@
-export const { router } = jest.requireActual("../router") as typeof import("../router");
-export const getCurrentUI = jest.fn();
