--- src/plugin/ui/index.ts
@@ -1,4 +1,0 @@
-export { ui, type UIController } from "./controller";
-export { type MessageHandler, type MessageRequest } from "./message";
-export { type PropertyInspector } from "./property-inspector";
-export { route } from "./route";
